requirements.txt
paramiko~=3.5.1
pyyaml~=6.0.2
requests~=2.32.3
google-cloud-firestore
pip~=23.2.1
wheel~=0.41.2
google~=3.0.0
protobuf~=5.29.3
cryptography~=44.0.2
setuptools~=68.2.0